+++ lib/AxisAlignedBB.h
@@ -10,12 +10,12 @@ class RayTraceResult; // forward declare to avoid include
 
 class AxisAlignedBB {
 public:
-    float minX = 0;
-    float minY = 0;
-    float minZ = 0;
-    float maxX = 0;
-    float maxY = 0;
-    float maxZ = 0;
+    double minX = 0;
+    double minY = 0;
+    double minZ = 0;
+    double maxX = 0;
+    double maxY = 0;
+    double maxZ = 0;
 
     /**
      * AxisAlignedBB constructor.
@@ -27,7 +27,7 @@ public:
      * @param maxY
      * @param maxZ
      */
-    AxisAlignedBB(float minX, float minY, float minZ, float maxX, float maxY, float maxZ) {
+    AxisAlignedBB(double minX, double minY, double minZ, double maxX, double maxY, double maxZ) {
         setBounds(minX, minY, minZ, maxX, maxY, maxZ);
     }
 
@@ -50,7 +50,7 @@ public:
      *
      * @return this
      */
-    AxisAlignedBB *setBounds(float minX, float minY, float minZ, float maxX, float maxY, float maxZ);
+    AxisAlignedBB *setBounds(double minX, double minY, double minZ, double maxX, double maxY, double maxZ);
 
     /**
      * Sets the bounding box's bounds from another AxisAlignedBB, and returns itself.
@@ -74,7 +74,7 @@ public:
      *
      * @return New modified bb.
      */
-    AxisAlignedBB *addCoord(float x, float y, float z);
+    AxisAlignedBB *addCoord(double x, double y, double z);
 
     /**
      * Outsets the bounds of this AxisAlignedBB by the specified X, Y and Z.
@@ -84,7 +84,7 @@ public:
      *
      * @return this
      */
-    AxisAlignedBB *expand(float x, float y, float z) {
+    AxisAlignedBB *expand(double x, double y, double z) {
         minX -= x;
         minY -= y;
         minZ -= z;
@@ -104,7 +104,7 @@ public:
      *
      * @return New modified bb.
      */
-    AxisAlignedBB *expandedCopy(float x, float y, float z) {
+    AxisAlignedBB *expandedCopy(double x, double y, double z) {
         return (new AxisAlignedBB(this))->expand(x, y, z);
     }
 
@@ -117,7 +117,7 @@ public:
      *
      * @return this
      */
-    AxisAlignedBB *offset(float x, float y, float z) {
+    AxisAlignedBB *offset(double x, double y, double z) {
         minX += x;
         minY += y;
         minZ += z;
@@ -137,7 +137,7 @@ public:
      *
      * @return New modified bb.
      */
-    AxisAlignedBB *offsetCopy(float x, float y, float z) {
+    AxisAlignedBB *offsetCopy(double x, double y, double z) {
         return (new AxisAlignedBB(this))->offset(x, y, z);
     }
 
@@ -150,7 +150,7 @@ public:
      *
      * @return this
      */
-    AxisAlignedBB *contract(float x, float y, float z) {
+    AxisAlignedBB *contract(double x, double y, double z) {
         minX += x;
         minY += y;
         minZ += z;
@@ -170,7 +170,7 @@ public:
      *
      * @return New modified bb.
      */
-    AxisAlignedBB *contractCopy(float x, float y, float z) {
+    AxisAlignedBB *contractCopy(double x, double y, double z) {
         return (new AxisAlignedBB(this))->contract(x, y, z);
     }
 
@@ -180,7 +180,7 @@ public:
      *
      * @return
      */
-    float calculateXOffset(const AxisAlignedBB *bb, float x);
+    double calculateXOffset(const AxisAlignedBB *bb, double x);
 
     /**
      * @param bb
@@ -188,7 +188,7 @@ public:
      *
      * @return
      */
-    float calculateYOffset(const AxisAlignedBB *bb, float y);
+    double calculateYOffset(const AxisAlignedBB *bb, double y);
 
     /**
      * @param bb
@@ -196,7 +196,7 @@ public:
      *
      * @return
      */
-    float calculateZOffset(const AxisAlignedBB *bb, float z);
+    double calculateZOffset(const AxisAlignedBB *bb, double z);
 
     /**
      * Returns whether any part of the specified AABB is inside (intersects with) this one.
@@ -206,7 +206,7 @@ public:
      *
      * @return
      */
-    bool intersectsWith(const AxisAlignedBB *bb, float epsilon = 0.00001);
+    bool intersectsWith(const AxisAlignedBB *bb, double epsilon = 0.00001);
 
     /**
      * Returns whether the specified vector is within the bounds of this AABB on all axes.
@@ -222,7 +222,7 @@ public:
      *
      * @return
      */
-    float getAverageEdgeLength() {
+    double getAverageEdgeLength() {
         return (maxX - minX + maxY - minY + maxZ - minZ) / 3;
     }
